+++ src/wiremind_kubernetes/kubernetes_helper.py
@@ -2,6 +2,7 @@
 import logging
 import os
 import time
+from typing import Any, Dict, List, Union
 
 import kubernetes
 
@@ -9,7 +10,6 @@ from wiremind_kubernetes.exceptions import PodNotFound
 
 from .kube_config import load_kubernetes_config
 from .utils import retry_kubernetes_request
-
 
 logger = logging.getLogger(__name__)
 
@@ -19,75 +19,68 @@ class KubernetesHelper:
     A simple helper for Kubernetes manipulation.
     """
 
-    client_corev1_api = None
-    client_batchv1_api = None
-    client_appsv1_api = None
-    client_custom_objects_api = None
-
-    def __init__(self, use_kubeconfig=False, dry_run=False):
+    def __init__(self, use_kubeconfig: bool = False, dry_run: bool = False):
         """
         :param use_kubeconfig:
             Use ~/.kube/config file to authenticate.
             If false, use kubernetes built-in in_cluster mechanism.
             Defaults to False.
-        :param deployment_namespace:
-            Target namespace to use.
-            If not defined, try to get it from kubernetes built-in serviceAccount mechanism.
-        :namespaced
-            True by default. If False, no specific namespace will be set by default.
-            Not all fonctions are compatible with non-namespaced.
         :param dry_run:
             Dry run.
         """
         load_kubernetes_config(use_kubeconfig=use_kubeconfig)
-        self.client_appsv1_api = kubernetes.client.AppsV1Api()
-        self.client_corev1_api = kubernetes.client.CoreV1Api()
-        self.client_batchv1_api = kubernetes.client.BatchV1Api()
-        self.client_custom_objects_api = kubernetes.client.CustomObjectsApi()
+        self.client_appsv1_api: kubernetes.client.AppsV1Api = kubernetes.client.AppsV1Api()
+        self.client_corev1_api: kubernetes.client.CoreV1Api = kubernetes.client.CoreV1Api()
+        self.client_batchv1_api: kubernetes.client.BatchV1Api = kubernetes.client.BatchV1Api()
+        self.client_custom_objects_api: kubernetes.client.CustomObjectsApi = kubernetes.client.CustomObjectsApi()
 
-        self.dry_run = dry_run
+        self.dry_run: bool = dry_run
 
         # Every read request have those arguments added
-        self.read_additional_arguments = dict(pretty=True)
+        self.read_additional_arguments: Dict[str, Any] = dict(pretty=True)
         # Every request, either read or write, have those arguments added
-        self.additional_arguments = self.read_additional_arguments.copy()
+        self.additional_arguments: Dict[
+            str, Any
+        ] = self.read_additional_arguments.copy()
         if dry_run:
             # Dry run, in kube API, is not true or false, but either dry_run: All or not defined.
             self.additional_arguments["dry_run"] = "All"
 
-    def delete_resource(self, resource_name, resource_namespace, delete_function):
+    def delete_resource(
+        self, resource_name: str, resource_namespace: str, delete_function
+    ):
         return delete_function(
             resource_name, resource_namespace, **self.additional_arguments
         )
 
-    def delete_pod(self, pod_name, pod_namespace):
+    def delete_pod(self, pod_name: str, pod_namespace: str):
         return self.delete_resource(
             pod_name, pod_namespace, self.client_corev1_api.delete_namespaced_pod
         )
 
-    def delete_pvc(self, pvc_name, pvc_namespace):
+    def delete_pvc(self, pvc_name: str, pvc_namespace: str):
         return self.delete_resource(
             pvc_name,
             pvc_namespace,
             self.client_corev1_api.delete_namespaced_persistent_volume_claim,
         )
 
-    def delete_job(self, job_name, job_namespace):
+    def delete_job(self, job_name: str, job_namespace: str):
         return self.delete_resource(
             job_name, job_namespace, self.client_batchv1_api.delete_namespaced_job
         )
 
-    def delete_secret(self, secret_name, secret_namespace):
+    def delete_secret(self, secret_name: str, secret_namespace: str):
         return self.delete_resource(
             secret_name,
             secret_namespace,
             self.client_corev1_api.delete_namespaced_secret,
         )
 
-    def delete_deployment(self, deployment_name, deployment_namespace):
+    def delete_deployment(self, deployment_name: str, namespace: str):
         return self.delete_resource(
             deployment_name,
-            deployment_namespace,
+            namespace,
             self.client_appsv1_api.delete_namespaced_deployment,
         )
 
@@ -122,127 +115,119 @@ class NamespacedKubernetesHelper(KubernetesHelper):
     A simple helper for Kubernetes manipulation.
     """
 
-    deployment_namespace = None
-
-    def __init__(self, use_kubeconfig=False, deployment_namespace=None, dry_run=False):
+    def __init__(
+        self,
+        use_kubeconfig: bool = False,
+        namespace: Union[None, str] = None,
+        dry_run: bool = False,
+    ):
         """
         :param use_kubeconfig:
             Use ~/.kube/config file to authenticate.
             If false, use kubernetes built-in in_cluster mechanism.
             Defaults to False.
-        :param deployment_namespace:
+        :param namespace:
             Target namespace to use.
             If not defined, try to get it from kubernetes built-in serviceAccount mechanism.
         :param dry_run:
             Dry run.
         """
         super().__init__(use_kubeconfig, dry_run)
-        if deployment_namespace:
-            self.deployment_namespace = deployment_namespace
+        if namespace:
+            self.namespace = namespace
         else:
-            self.deployment_namespace = open(
+            self.namespace = open(
                 "/var/run/secrets/kubernetes.io/serviceaccount/namespace"
             ).read()
 
-    def get_deployment_scale(self, deployment_name):
+    def get_deployment_scale(self, deployment_name: str):
         logger.debug("Getting deployment scale for %s", deployment_name)
         return self.client_appsv1_api.read_namespaced_deployment_scale(
-            deployment_name, self.deployment_namespace, **self.read_additional_arguments
+            deployment_name, self.namespace, **self.read_additional_arguments
         )
 
-    def get_statefulset_scale(self, statefulset_name):
+    def get_statefulset_scale(self, statefulset_name: str):
         logger.debug("Getting statefulset scale for %s", statefulset_name)
         return self.client_appsv1_api.read_namespaced_stateful_set_scale(
-            statefulset_name,
-            self.deployment_namespace,
-            **self.read_additional_arguments,
+            statefulset_name, self.namespace, **self.read_additional_arguments,
         )
 
-    def scale_down_statefulset(self, statefulset_name):
+    def scale_down_statefulset(self, statefulset_name: str):
         body = self.get_statefulset_scale(statefulset_name)
         logger.debug("Deleting all Pods for %s", statefulset_name)
         body.spec.replicas = 0
         self.client_appsv1_api.patch_namespaced_stateful_set_scale(
-            statefulset_name,
-            self.deployment_namespace,
-            body,
-            **self.additional_arguments,
+            statefulset_name, self.namespace, body, **self.additional_arguments,
         )
         logger.debug("Done deleting.")
 
     @retry_kubernetes_request
-    def scale_down_deployment(self, deployment_name):
+    def scale_down_deployment(self, deployment_name: str):
         body = self.get_deployment_scale(deployment_name)
         logger.debug("Deleting all Pods for %s", deployment_name)
         body.spec.replicas = 0
         self.client_appsv1_api.patch_namespaced_deployment_scale(
-            deployment_name,
-            self.deployment_namespace,
-            body,
-            **self.additional_arguments,
+            deployment_name, self.namespace, body, **self.additional_arguments,
         )
         logger.debug("Done deleting.")
 
-    def scale_up_statefulset(self, statefulset_name, pod_amount=1):
+    def scale_up_statefulset(self, statefulset_name: str, pod_amount: int = 1):
         body = self.get_statefulset_scale(statefulset_name)
         logger.debug("Recreating backend Pods for %s", statefulset_name)
         body.spec.replicas = pod_amount
         self.client_appsv1_api.patch_namespaced_stateful_set_scale(
-            statefulset_name,
-            self.deployment_namespace,
-            body,
-            **self.additional_arguments,
+            statefulset_name, self.namespace, body, **self.additional_arguments,
         )
         logger.debug("Done recreating.")
 
     @retry_kubernetes_request
-    def scale_up_deployment(self, deployment_name, pod_amount):
+    def scale_up_deployment(self, deployment_name: str, pod_amount: int):
         body = self.get_deployment_scale(deployment_name)
         logger.debug("Recreating backend Pods for %s", deployment_name)
         body.spec.replicas = pod_amount
         self.client_appsv1_api.patch_namespaced_deployment_scale(
-            deployment_name,
-            self.deployment_namespace,
-            body,
-            **self.additional_arguments,
+            deployment_name, self.namespace, body, **self.additional_arguments,
         )
         logger.debug("Done recreating.")
 
-    def is_statefulset_stopped(self, deployment_name):
+    def is_statefulset_stopped(self, deployment_name: str):
         return self.is_deployment_stopped(deployment_name, statefulset=True)
 
     @retry_kubernetes_request
-    def is_deployment_stopped(self, deployment_name, statefulset=False):
+    def is_deployment_stopped(self, deployment_name: str, statefulset: bool = False):
         logger.debug("Asking if deployment %s is stopped", deployment_name)
         if statefulset:
             scale = self.client_appsv1_api.read_namespaced_stateful_set_scale(
-                deployment_name,
-                self.deployment_namespace,
-                **self.read_additional_arguments,
+                deployment_name, self.namespace, **self.read_additional_arguments,
             )
         else:
             scale = self.client_appsv1_api.read_namespaced_deployment_scale(
-                deployment_name,
-                self.deployment_namespace,
-                **self.read_additional_arguments,
+                deployment_name, self.namespace, **self.read_additional_arguments,
             )
         logger.debug("%s has %s replicas", deployment_name, scale.status.replicas)
-        return scale.status.replicas == 0 or self.dry_run
+        return (scale.status.replicas == 0) or self.dry_run
 
-    def is_statefulset_ready(self, statefulset_name):
-        statefulset_status = self.client_appsv1_api.read_namespaced_stateful_set_status(
-            statefulset_name,
-            self.deployment_namespace,
-            **self.read_additional_arguments,
-        )
-        expected_replicas = statefulset_status.spec.replicas
-        ready_replicas = statefulset_status.status.ready_replicas
+    def is_deployment_ready(self, deployment_name: str, statefulset: bool = False):
+        if statefulset:
+            status = self.client_appsv1_api.read_namespaced_stateful_set_status(
+                deployment_name, self.namespace, **self.read_additional_arguments,
+            )
+        else:
+            status = self.client_appsv1_api.read_namespaced_deployment_status(
+                deployment_name, self.namespace, **self.read_additional_arguments,
+            )
+        expected_replicas = status.spec.replicas
+        ready_replicas = status.status.ready_replicas
+        resource_type = statefulset and "StatefulSet" or "Deployment"
         logger.debug(
-            "StatefulSet %s has %s expected replicas and %s ready replicas",
-            statefulset_name,
+            "%s %s has %s expected replicas and %s ready replicas",
+            resource_type,
+            deployment_name,
             expected_replicas,
             ready_replicas,
         )
+        if expected_replicas == 0:
+            return True
         return expected_replicas == ready_replicas
 
     @retry_kubernetes_request
@@ -284,7 +269,7 @@ class KubernetesDeploymentManager(NamespacedKubernetesHelper):
     The associated Deployment should define an eds.
 
     Usage:
-    a = wiremind_kubernetes.KubernetesDeploymentManager(use_kubeconfig=True, deployment_namespace="my-namespace")
+    a = wiremind_kubernetes.KubernetesDeploymentManager(use_kubeconfig=True, namespace="my-namespace")
     a.stop_pods()
     do_something('wololo')
     a.start_pods()
@@ -295,7 +280,7 @@ class KubernetesDeploymentManager(NamespacedKubernetesHelper):
             self.release_name = release_name
         else:
             self.release_name = os.environ.get("RELEASE_NAME")
-        super(KubernetesDeploymentManager, self).__init__(**kwargs)
+        super().__init__(**kwargs)
 
     def start_pods(self):
         """
@@ -351,14 +336,14 @@ class KubernetesDeploymentManager(NamespacedKubernetesHelper):
         logger.debug("Getting Expected Deployment Scale list")
         if not self.release_name:
             eds_list = self.client_custom_objects_api.list_namespaced_custom_object(
-                namespace=self.deployment_namespace,
+                namespace=self.namespace,
                 group="wiremind.fr",
                 version="v1",
                 plural="expecteddeploymentscales",
             )
         else:
             eds_list = self.client_custom_objects_api.list_namespaced_custom_object(
-                namespace=self.deployment_namespace,
+                namespace=self.namespace,
                 group="wiremind.fr",
                 version="v1",
                 plural="expecteddeploymentscales",
@@ -366,7 +351,7 @@ class KubernetesDeploymentManager(NamespacedKubernetesHelper):
             )
             # Support new-style labels as well
             additional_eds_list = self.client_custom_objects_api.list_namespaced_custom_object(
-                namespace=self.deployment_namespace,
+                namespace=self.namespace,
                 group="wiremind.fr",
                 version="v1",
                 plural="expecteddeploymentscales",
@@ -379,3 +364,49 @@ class KubernetesDeploymentManager(NamespacedKubernetesHelper):
         }
         logger.debug("List is %s", eds_dict)
         return eds_dict
+
+    def create_job(
+        self,
+        job_name: str,
+        container_image: str,
+        command: Union[str, None] = None,
+        args: Union[List[str], None] = None,
+        environment_variables: Union[Dict["str", "str"], None] = None,
+        ttl_seconds_after_finished: int = 1800,
+    ):
+        job_name = f"{self.release_name}-{job_name}"
+        if environment_variables is None:
+            environment_variables = {}
+
+        body = kubernetes.client.V1Job(api_version="batch/v1", kind="Job")
+        body.metadata = kubernetes.client.V1ObjectMeta(
+            namespace=self.namespace, name=job_name
+        )
+        body.status = kubernetes.client.V1JobStatus()
+        template: kubernetes.client.V1PodTemplate = kubernetes.client.V1PodTemplate()
+        template.template = kubernetes.client.V1PodTemplateSpec()
+        env_list = []
+        for env_name, env_value in environment_variables.items():
+            env_list.append(kubernetes.client.V1EnvVar(name=env_name, value=env_value))
+        container = kubernetes.client.V1Container(
+            name="wiremind", image=container_image, env=env_list
+        )
+        if command:
+            container.command = [command]
+        if args:
+            container.args = args
+        template.template.spec = kubernetes.client.V1PodSpec(
+            containers=[container], restart_policy="Never"
+        )
+        body.spec = kubernetes.client.V1JobSpec(
+            ttl_seconds_after_finished=ttl_seconds_after_finished,
+            template=template.template,
+        )
+
+        try:
+            self.client_batchv1_api.create_namespaced_job(
+                self.namespace, body, **self.additional_arguments
+            )
+        except kubernetes.client.rest.ApiException as e:
+            print("Exception when calling BatchV1Api->create_namespaced_job: %s\n" % e)
+        return